# docs/swagger.yaml
basePath: /
definitions:
  main.Airport:
    properties:
      airport_code:
        type: string
      airport_name:
        type: string
      city:
        type: string
      timezone:
        type: string
    type: object
  main.BoardingPass:
    properties:
      boarding_no:
        type: integer
      flight_id:
        type: integer
      seat_no:
        type: string
      ticket_no:
        type: string
    type: object
  main.BookingRequest:
    properties:
      fare_conditions:
        type: string
      flight_ids:
        items:
          type: integer
        type: array
      passanger:
        type: string
    type: object
  main.FlightSchedule:
    properties:
      day_of_week:
        type: string
      flight_no:
        type: string
      origin_airport:
        type: string
      time_of_arrival:
        type: string
    type: object
  main.TicketFlight:
    properties:
      fare_conditions:
        type: string
      flight_id:
        type: integer
      ticket_no:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  title: Airflight Service API
  version: "1.0"
paths:
  /airports:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all airports from the database
      parameters:
      - description: Airport city
        in: query
        name: city
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Airport'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all airports
      tags:
      - airports
  /airports/{airport_code}/inbound-schedule:
    get:
      description: Retrieves the inbound flight schedule for a specified airport
      parameters:
      - description: Airport code
        in: path
        name: airport_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.FlightSchedule'
            type: array
        "400":
          description: Missing or invalid airport code
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get inbound schedule for an airport
      tags:
      - airports
  /airports/{airport_code}/outbound-schedule:
    get:
      description: Retrieves the outbound flight schedule for a specified airport
      parameters:
      - description: Airport code
        in: path
        name: airport_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.FlightSchedule'
            type: array
        "400":
          description: Missing or invalid airport code
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get outbound schedule for an airport
      tags:
      - airports
  /bookings/{guid}:
    put:
      consumes:
      - application/json
      description: Idempotent booking of flights with a GUID
      parameters:
      - description: GUID
        in: path
        name: guid
        required: true
        type: string
      - description: Booking data
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/main.BookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Existing or new tickets
          schema:
            items:
              $ref: '#/definitions/main.TicketFlight'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Book a route
      tags:
      - bookings
  /bookings/{guid}/check-in/{flight_id}:
    put:
      consumes:
      - application/json
      description: Assigns a seat for a booked flight using a GUID
      parameters:
      - description: Booking GUID
        in: path
        name: guid
        required: true
        type: string
      - description: Flight ID
        in: path
        name: flight_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Boarding pass details
          schema:
            $ref: '#/definitions/main.BoardingPass'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Booking or seat not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Check-in for a flight
      tags:
      - bookings
  /cities:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all cities from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all cities
      tags:
      - cities
swagger: "2.0"